# doc/openapi.yaml
openapi: 3.0.3
info:
  title: 应用发布服务 API
  version: 1.0.0
  description: 当前应用发布服务接口的 OpenAPI 定义。
servers:
  - url: http://localhost:8080
    description: 本地环境
...
paths:
  /api/apps:
    post:
      tags: [应用管理]
      summary: 创建应用
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAppResponse'
    get:
      tags: [应用管理]
      summary: 查询应用列表
      parameters:
        - in: query
          name: keyword
          description: 应用名称或包名关键字
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAppResponseList'
  /api/apps/{appId}:
    get:
      tags: [应用管理]
      summary: 查询应用详情
      parameters:
        - $ref: '#/components/parameters/AppId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAppDetailResponse'
    put:
      tags: [应用管理]
      summary: 更新应用
      parameters:
        - $ref: '#/components/parameters/AppId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAppResponse'
    delete:
      tags: [应用管理]
      summary: 删除应用
      parameters:
        - $ref: '#/components/parameters/AppId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
  /api/apps/{appId}/versions:
    get:
      tags: [安装包管理]
      summary: 查询版本列表
      parameters:
        - $ref: '#/components/parameters/AppId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAppVersionResponseList'
  /api/apps/{appId}/versions/{versionId}:
    get:
      tags: [安装包管理]
      summary: 查询版本详情
      parameters:
        - $ref: '#/components/parameters/AppId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAppVersionResponse'
  /api/apps/{appId}/versions/upload:
    post:
      tags: [安装包管理]
      summary: 上传安装包
      parameters:
        - $ref: '#/components/parameters/AppId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                updateLog:
                  type: string
                  description: 更新说明
                expectedVersionName:
                  type: string
                  description: 期望解析出的版本名称
                expectedVersionCode:
                  type: integer
                  description: 期望解析出的版本号
                expectedReinforced:
                  type: boolean
                  description: 期望解析出的是否已加固标记
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePackageUploadResponse'
  /api/releases:
    get:
      tags: [发版管理]
      summary: 查询发版记录
      parameters:
        - in: query
          name: current
          description: 当前页，兼容 pageNum
          schema:
            type: integer
            format: int64
        - in: query
          name: size
          description: 每页大小，兼容 pageSize
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNum
          description: 页码
          schema:
            type: integer
            format: int64
        - in: query
          name: pageSize
          description: 每页条数
          schema:
            type: integer
            format: int64
        - in: query
          name: key
          description: 关键字
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageResponseReleaseRecordPageResponse'
  /api/releases/appId/{appId}:
    get:
      tags: [发版管理]
      summary: 根据应用 ID 查询发版记录
      parameters:
        - $ref: '#/components/parameters/AppId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageResponseReleaseRecordPageResponse'
  /api/releases/submit:
    post:
      tags: [发版管理]
      summary: 提交发版
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseSubmitRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReleaseRecordResponseList'
  /api/releases/{releaseId}:
    get:
      tags: [发版管理]
      summary: 查询发版详情
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReleaseRecordResponse'
  /api/releases/{releaseId}/logs:
    get:
      tags: [发版管理]
      summary: 查询发版日志
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReleaseTaskLogResponseList'
...
components:
  parameters:
    AppId:
      in: path
      name: appId
      description: 应用 ID
      required: true
      schema:
        type: integer
        format: int64
...
    VersionId:
      in: path
      name: versionId
      description: 版本 ID
      required: true
      schema:
        type: integer
        format: int64
    ReleaseId:
      in: path
      name: releaseId
      description: 发版记录 ID
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    AppUpsertRequest:
      type: object
      required: [appName, packageName, appType]
      properties:
        id:
          type: integer
          format: int64
        appName:
          type: string
        packageName:
          type: string
        appType:
          type: integer
          enum: [1, 2]
          description: 应用类型，1=Android 应用，2=iOS 应用
        appDescription:
          type: string
        copyrightNo:
          type: string
        icpNo:
          type: string
        appRecordNo:
          type: string
        privacyUrl:
          type: string
        userAgreementUrl:
          type: string
        status:
          type: integer
          enum: [0, 1]
          description: 应用状态，0=禁用，1=启用
        versionCode:
          type: integer
          description: 版本号，创建应用时传入会自动初始化一条版本记录
...
    ReleaseSubmitRequest:
      type: object
      required: [versionId, storeTypes]
      properties:
        versionId:
          type: integer
          format: int64
          description: 应用版本 ID
        storeTypes:
          type: array
          description: 发版渠道类型编码列表
          items:
            type: string
        releaseMode:
          type: string
          default: api
          enum: [manual, api]
          description: 发版模式，manual=人工发版，api=接口发版
        releaseType:
          type: integer
          format: int64
          default: 1
          description: 1=全量发布 2=分阶段发布
        grayPercent:
          type: integer
          format: int64
          description: 分阶段灰度比例，releaseType=2 时必填
        grayStartTime:
          type: string
          format: date-time
          description: 分阶段开始时间，releaseType=2 时必填
        grayEndTime:
          type: string
          format: date-time
          description: 分阶段结束时间，releaseType=2 时必填
    AppResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        appName:
          type: string
        packageName:
          type: string
        appType:
          type: integer
          description: 应用类型，1=Android 应用，2=iOS 应用
        appDescription:
          type: string
        copyrightNo:
          type: string
        icpNo:
          type: string
        appRecordNo:
          type: string
        privacyUrl:
          type: string
        userAgreementUrl:
          type: string
        status:
          type: integer
          description: 状态，1=启用，0=禁用
        createUser:
          type: string
        updateUser:
          type: string
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
    AppDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        appName:
          type: string
        packageName:
          type: string
        appType:
          type: integer
          description: 应用类型，1=Android 应用，2=iOS 应用
        appDescription:
          type: string
        copyrightNo:
          type: string
        icpNo:
          type: string
        appRecordNo:
          type: string
        privacyUrl:
          type: string
        userAgreementUrl:
          type: string
        status:
          type: integer
          description: 状态，1=启用，0=禁用
        createUser:
          type: string
        updateUser:
          type: string
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
        versions:
          type: array
          description: 应用版本列表
          items:
            $ref: '#/components/schemas/AppVersionResponse'
        releaseRecords:
          type: array
          description: 发版记录列表
          items:
            $ref: '#/components/schemas/ReleaseRecordResponse'
...
    AppVersionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        appId:
          type: integer
          format: int64
        versionName:
          type: string
          description: 版本名称
        versionCode:
          type: integer
          description: 版本号
        packageUrl:
          type: string
          description: 安装包访问地址
        updateLog:
          type: string
          description: 更新说明
        reinforced:
          type: boolean
          description: 是否为加固包
        createUser:
          type: string
        updateUser:
          type: string
        createTime:
          type: string
          format: date-time
    PackageUploadResponse:
      type: object
      properties:
        versionId:
          type: integer
          format: int64
        appId:
          type: integer
          format: int64
        packageType:
          type: string
          description: 安装包类型
        storedPath:
          type: string
          description: 文件存储路径
        versionName:
          type: string
        versionCode:
          type: integer
          description: 版本号
        reinforced:
          type: boolean
          description: 是否已加固
        checksum:
          type: string
          description: 文件校验值
        createTime:
          type: string
          format: date-time
          description: 上传时间
    ReleaseRecordResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        appId:
          type: integer
          format: int64
        appName:
          type: string
          description: 应用名称
        packageName:
          type: string
          description: 应用包名
        appDescription:
          type: string
          description: 应用描述
        versionId:
          type: integer
          format: int64
        versionCode:
          type: integer
          description: 版本号
        storeType:
          type: string
          description: 商店类型
        releaseMode:
          type: string
          description: 发布方式
        releaseType:
          type: integer
          format: int64
          description: 发布类型，1全量 2分阶段发布
        grayPercent:
          type: integer
          format: int64
          description: 灰度比例
        grayStartTime:
          type: string
          format: date-time
          description: 分阶段开始时间
        grayEndTime:
          type: string
          format: date-time
          description: 分阶段结束时间
        releaseStatus:
          type: string
          description: 发布状态
        storeReleaseId:
          type: string
          description: 渠道侧发版单号
        rejectReason:
          type: string
          description: 驳回原因
        apiRequestLog:
          type: string
          description: 接口请求日志
        apiResponseLog:
          type: string
          description: 接口响应日志
        releaseTime:
          type: string
          format: date-time
          description: 提审时间
        finishTime:
          type: string
          format: date-time
          description: 发版完成时间
        createUser:
          type: string
        updateUser:
          type: string
    ReleaseRecordPageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        appId:
          type: integer
          format: int64
        appName:
          type: string
          description: 应用名称
        packageName:
          type: string
          description: 应用包名
        appDescription:
          type: string
          description: 应用描述
        versionId:
          type: integer
          format: int64
        versionCode:
          type: integer
          description: 版本号
        storeType:
          type: string
          description: 商店类型
        releaseMode:
          type: string
          description: 发布方式
        releaseStatus:
          type: string
          description: 发布状态
        releaseTime:
          type: string
          format: date-time
          description: 提审时间
        finishTime:
          type: string
          format: date-time
          description: 发版完成时间
        createUser:
          type: string
        updateUser:
          type: string
    PageResponseReleaseRecordPageResponse:
      type: object
      properties:
        current:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
        pages:
          type: integer
          format: int64
        records:
          type: array
          items:
            $ref: '#/components/schemas/ReleaseRecordPageResponse'
    ReleaseTaskLogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        releaseRecordId:
          type: integer
          format: int64
        action:
          type: string
          description: 执行动作
        statusBefore:
          type: string
          description: 变更前状态
        statusAfter:
          type: string
          description: 变更后状态
        message:
          type: string
          description: 日志消息
        payload:
          type: string
          description: 附加上下文数据
        createUser:
          type: string
        updateUser:
          type: string
        createTime:
          type: string
          format: date-time
    ApiResponseAppResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AppResponse'
        message:
          type: string
    ApiResponseAppDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AppDetailResponse'
        message:
          type: string
    ApiResponseAppResponseList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/AppResponse'
        message:
          type: string
...
    ApiResponseAppVersionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AppVersionResponse'
        message:
          type: string
    ApiResponseAppVersionResponseList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/AppVersionResponse'
        message:
          type: string
    ApiResponsePackageUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PackageUploadResponse'
        message:
          type: string
    ApiResponseReleaseRecordResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ReleaseRecordResponse'
        message:
          type: string
    ApiResponseReleaseRecordResponseList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReleaseRecordResponse'
        message:
          type: string
    ApiResponsePageResponseReleaseRecordPageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PageResponseReleaseRecordPageResponse'
        message:
          type: string
    ApiResponseReleaseTaskLogResponseList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReleaseTaskLogResponse'
        message:
          type: string
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string