# src/main/java/com/app/publishservice/api/dto/AppVersionResponse.java
package com.app.publishservice.api.dto;

import io.swagger.v3.oas.annotations.media.Schema;

import java.time.LocalDateTime;

@Schema(description = "应用版本响应")
public record AppVersionResponse(
        @Schema(description = "版本 ID") Long id,
        @Schema(description = "应用 ID") Long appId,
        @Schema(description = "版本名称") String versionName,
        @Schema(description = "版本号") Integer versionCode,
        @Schema(description = "安装包访问地址") String packageUrl,
        @Schema(description = "更新说明") String updateLog,
        @Schema(description = "是否为加固包") boolean reinforced,
        @Schema(description = "创建人") String createUser,
        @Schema(description = "更新人") String updateUser,
        @Schema(description = "创建时间") LocalDateTime createTime
) {
}